new changes

+++ include/utils/radix.h
@@ -40,8 +40,8 @@ std::unique_ptr<size_t[]> idx_radixsort(float const* fvalues,
  *  @param nvalues length of \a fvalues
  *  @return indexes of ascending sorted \a fvalues
  */
-size_t *idxfloat_radixsort(float const* fvalues,
-                                 const size_t nvalues);
+unsigned int *idxfloat_radixsort(float const* fvalues,
+                                 const unsigned int nvalues);
 
 enum sortorder {
   ascending,
@@ -53,7 +53,7 @@ enum sortorder {
  *  @param nvalues length of \a fvalues
  */
 template<sortorder const order> void float_radixsort(
-    float *fvalues, const size_t nvalues);
+    float *fvalues, const unsigned int nvalues);
 
 /*! sort an array of float values with respect to another one without modifing the input array and returning permuted indexes of the sorted items
  *  @param extvalues input float array
@@ -62,7 +62,7 @@ template<sortorder const order> void float_radixsort(
  *  @return a sorted copy of \a extvalues wrt \a fvalues
  */
 template<sortorder const order> float *copyextfloat_radixsort(
-    float const* extvalues, float const* fvalues, const size_t nvalues);
+    float const* extvalues, float const* fvalues, const unsigned int nvalues);
 
 #endif
